order-chatbot/requirements.txt
langchain>=0.1.0
langchain-community
google-generativeai>=0.3.0
google-cloud-aiplatform
mysql-connector-python
python-dotenv
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pydantic
python-multipart
requests
typing-extensions
aiohttp
openai
tiktoken
numpy
pandas
sqlalchemy
pytest
black
flake8
mypy
google-generativeai
